+++ src/api_server/impl/SubscriptionCreationApiImpl.cpp
@@ -54,7 +54,7 @@ void SubscriptionCreationApiImpl::subscribe(
     const std::string& supi, const SdmSubscription& sdmSubscription,
     Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
-  long http_code               = {};
+  uint32_t http_code           = {};
   Pistache::Http::Code code    = {};
 
   m_udm_app->handle_subscription_creation(
